Try to fix failed deployment (more details inside)
deploy failed because the charterEdits table isn't called
charteredits... was fine on my machine using sqlite bc ig table names
are case-insensitive in sqlite but they're not in postgres. Anyway this
is all a wild guess so if you're reading this and I'm wrong feel free to
point and laugh.